requirements.txt
Flask
Flask-Autodoc
Flask-Login
Flask-Restful
Flask-Sqlalchemy
Flask-Testing
Flask-WTF
WTForms-Alchemy
datadog
datadog_logger
dateutils
flask_bootstrap4
flask_debugtoolbar
gunicorn
jinja2
marshmallow
passlib
praw
psycopg2-binary
sentry-sdk[flask]
sphinx
sphinx_rtd_theme
sphinxcontrib-httpdomain
sphinxcontrib-sadisplay
sqlalchemy
sshtunnel
werkzeug